+++ lib/language_pack/ruby.rb
@@ -99,6 +99,7 @@ WARNING
       setup_export
       setup_profiled
       allow_git do
+        vendor_libpq
         install_bundler_in_app
         build_bundler("development:test")
         post_bundler
@@ -108,8 +109,61 @@ WARNING
       end
       config_detect
       best_practice_warnings
+      warn_outdated_ruby
       cleanup
       super
+    end
+  rescue => e
+    warn_outdated_ruby
+    raise e
+  end
+
+  def vendor_libpq
+    # Check for existing libraries
+    return unless File.exist?("/usr/lib/x86_64-linux-gnu/libpq.so.5.11")
+    return unless ENV['STACK'] == 'heroku-18'
+
+    topic("Vendoring libpq 5.12.1")
+
+    @metadata.fetch("vendor_libpq12") do
+      warn(<<~EOF)
+        Replacing libpq with version libpq 5.12.1
+
+        This version includes a bug fix that can cause an exception
+        on boot for applications with incorrectly configured connection
+        values. For more information see:
+
+          https://devcenter.heroku.com/articles/libpq-5-12-1-breaking-connection-behavior
+
+        If your application breaks you can rollback to your last build.
+        You can also temporarially opt out of this behavior by setting:
+
+        ```
+        $ heroku config:set HEROKU_SKIP_LIBPQ12=1
+        ```
+
+        In the future libpq 5.12 will be the default on the platform and
+        you will not be able to opt-out of the library. For more information see:
+
+          https://devcenter.heroku.com/articles/libpq-5-12-1-breaking-connection-behavior
+      EOF
+
+      "true" # Set future cache value
+    end
+
+    Dir.chdir("vendor") do
+      @fetchers[:mri].fetch("libpq5_12.1-1.deb")
+      run!("dpkg -x libpq5_12.1-1.deb .")
+      run!("rm libpq5_12.1-1.deb")
+
+      load_libpq_12_unless_env_var = <<~EOF
+        if [ "$HEROKU_SKIP_LIBPQ12" == "" ]; then
+          export LD_LIBRARY_PATH="$HOME/vendor/usr/lib/x86_64-linux-gnu/:$LD_LIBRARY_PATH"
+        fi
+      EOF
+      add_to_export load_libpq_12_unless_env_var
+      add_to_profiled load_libpq_12_unless_env_var
+      ENV["LD_LIBRARY_PATH"] = Dir.pwd + "/usr/lib/x86_64-linux-gnu:#{ENV["LD_LIBRARY_PATH"]}"
     end
   end
 
@@ -121,16 +175,32 @@ WARNING
 
 private
 
+  def default_malloc_arena_max?
+    return true if @metadata.exists?("default_malloc_arena_max")
+    return @metadata.touch("default_malloc_arena_max") if new_app?
+
+    return false
+  end
+
+  def stack_not_14_not_16?
+    case stack
+    when "cedar-14", "heroku-16"
+      return false
+    else
+      return true
+    end
+  end
+
   def warn_bundler_upgrade
     old_bundler_version  = @metadata.read("bundler_version").chomp if @metadata.exists?("bundler_version")
 
     if old_bundler_version && old_bundler_version != bundler.version
-      puts(<<-WARNING)
+      warn(<<-WARNING, inline: true)
 Your app was upgraded to bundler #{ bundler.version }.
 Previously you had a successful deploy with bundler #{ old_bundler_version }.
 
 If you see problems related to the bundler version please refer to:
-https://devcenter.heroku.com/articles/bundler-version
+https://devcenter.heroku.com/articles/bundler-version#known-upgrade-issues
 
 WARNING
     end
@@ -219,23 +289,24 @@ WARNING
   # default JAVA_OPTS
   # return [String] string of JAVA_OPTS
   def default_java_opts
-    "-Xss512k -XX:+UseCompressedOops -Dfile.encoding=UTF-8"
+    "-Dfile.encoding=UTF-8"
   end
 
   def set_jvm_max_heap
     <<-EOF
-case $(ulimit -u) in
-256)   # 1X Dyno
-  JVM_MAX_HEAP=384
+limit=$(ulimit -u)
+case $limit in
+512)   # 2X, private-s: memory.limit_in_bytes=1073741824
+  echo "$opts -Xmx671m -XX:CICompilerCount=2"
   ;;
-512)   # 2X Dyno
-  JVM_MAX_HEAP=768
+16384) # perf-m, private-m: memory.limit_in_bytes=2684354560
+  echo "$opts -Xmx2g"
   ;;
-16384) # IX Dyno
-  JVM_MAX_HEAP=2048
+32768) # perf-l, private-l: memory.limit_in_bytes=15032385536
+  echo "$opts -Xmx12g"
   ;;
-32768) # PX Dyno
-  JVM_MAX_HEAP=5120
+*) # Free, Hobby, 1X: memory.limit_in_bytes=536870912
+  echo "$opts -Xmx300m -Xss512k -XX:CICompilerCount=2"
   ;;
 esac
 EOF
@@ -351,6 +422,7 @@ SHELL
       set_env_override "GEM_PATH", "$HOME/#{slug_vendor_base}:$GEM_PATH"
       set_env_override "PATH",      profiled_path.join(":")
 
+      set_env_default "MALLOC_ARENA_MAX", "2"     if default_malloc_arena_max?
       add_to_profiled set_default_web_concurrency if env("SENSIBLE_DEFAULTS")
 
       if ruby_version.jruby?
@@ -362,18 +434,95 @@ SHELL
     end
   end
 
+  def warn_outdated_ruby
+    return unless defined?(@outdated_version_check)
+
+    @warn_outdated ||= begin
+      @outdated_version_check.join
+
+      warn_outdated_minor
+      warn_outdated_eol
+      true
+    end
+  end
+
+  def warn_outdated_eol
+    return unless @outdated_version_check.maybe_eol?
+
+    if @outdated_version_check.eol?
+      warn(<<~WARNING)
+        EOL Ruby Version
+
+        You are using a Ruby version that has reached its End of Life (EOL)
+
+        We strongly suggest you upgrade to Ruby #{@outdated_version_check.suggest_ruby_eol_version} or later
+
+        Your current Ruby version no longer receives security updates from
+        Ruby Core and may have serious vulnerabilities. While you will continue
+        to be able to deploy on Heroku with this Ruby version you must upgrade
+        to a non-EOL version to be eligable to receive support.
+
+        Upgrade your Ruby version as soon as possible.
+
+        For a list of supported Ruby versions see:
+          https://devcenter.heroku.com/articles/ruby-support#supported-runtimes
+      WARNING
+    else
+      # Maybe EOL
+      warn(<<~WARNING)
+        Potential EOL Ruby Version
+
+        You are using a Ruby version that has either reached its End of Life (EOL)
+        or will reach its End of Life on December 25th of this year.
+
+        We suggest you upgrade to Ruby #{@outdated_version_check.suggest_ruby_eol_version} or later
+
+        Once a Ruby version becomes EOL, it will no longer receive
+        security updates from Ruby core and may have serious vulnerabilities.
+
+        Please upgrade your Ruby version.
+
+        For a list of supported Ruby versions see:
+          https://devcenter.heroku.com/articles/ruby-support#supported-runtimes
+      WARNING
+    end
+  end
+
+  def warn_outdated_minor
+    return if @outdated_version_check.latest_minor_version?
+
+    warn(<<~WARNING)
+      There is a more recent Ruby version available for you to use:
+
+      #{@outdated_version_check.suggested_ruby_minor_version}
+
+      The latest version will include security and bug fixes, we always recommend
+      running the latest version of your minor release.
+
+      Please upgrade your Ruby version.
+
+      For all available Ruby versions see:
+        https://devcenter.heroku.com/articles/ruby-support#supported-runtimes
+    WARNING
+  end
+
   # install the vendored ruby
   # @return [Boolean] true if it installs the vendored ruby and false otherwise
   def install_ruby
     instrument 'ruby.install_ruby' do
       return false unless ruby_version
-
       installer = LanguagePack::Installers::RubyInstaller.installer(ruby_version).new(@stack)
 
       if ruby_version.build?
         installer.fetch_unpack(ruby_version, build_ruby_path, true)
       end
       installer.install(ruby_version, slug_vendor_ruby)
+
+      @outdated_version_check = LanguagePack::Helpers::OutdatedRubyVersion.new(
+        current_ruby_version: ruby_version,
+        fetcher: installer.fetcher
+      )
+      @outdated_version_check.call
 
       @metadata.write("buildpack_ruby_version", ruby_version.version_for_download)
 
@@ -567,11 +716,7 @@ ERROR
   # install libyaml into the LP to be referenced for psych compilation
   # @param [String] tmpdir to store the libyaml files
   def install_libyaml(dir)
-    case stack
-    when "cedar-14", "heroku-16"
-    else
-      return
-    end
+    return false if stack_not_14_not_16?
 
     instrument 'ruby.install_libyaml' do
       FileUtils.mkdir_p dir
@@ -638,6 +783,12 @@ BUNDLE
   # runs bundler to install the dependencies
   def build_bundler(default_bundle_without)
     instrument 'ruby.build_bundler' do
+      topic("Removing BUNDLED WITH version in the Gemfile.lock")
+      contents = File.read("Gemfile.lock")
+      File.open("Gemfile.lock", "w") do |f|
+        f.write contents.sub(/^BUNDLED WITH$(\r?\n)   (?<major>\d+)\.\d+\.\d+/m, '')
+      end
+
       log("bundle") do
         bundle_without = env("BUNDLE_WITHOUT") || default_bundle_without
         bundle_bin     = "bundle"
@@ -1014,6 +1165,9 @@ params = CGI.parse(uri.query || "")
       bundler_version_cache   = "bundler_version"
       rubygems_version_cache  = "rubygems_version"
       stack_cache             = "stack"
+
+      # bundle clean does not remove binstubs
+      FileUtils.rm_rf("vendor/bundler/bin")
 
       old_rubygems_version = @metadata.read(ruby_version_cache).chomp if @metadata.exists?(ruby_version_cache)
       old_stack = @metadata.read(stack_cache).chomp if @metadata.exists?(stack_cache)